move tests cases into single test class for update readme tests
