+++ ui-tests-starter/tst-243+/software/aws/toolkits/jetbrains/uitests/docTests/updateReadmeTests/UpdateReadmeLatestChangesTest.kt
@@ -1,7 +1,7 @@
 // Copyright 2025 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package software.aws.toolkits.jetbrains.uitests.docTests.updateReadmeTests.updateReadmeLatestChanges
+package software.aws.toolkits.jetbrains.uitests.docTests.updateReadmeTests
 
 import com.intellij.driver.sdk.waitForProjectOpen
 import com.intellij.ide.starter.ci.CIServer
@@ -15,6 +15,7 @@ import com.intellij.ide.starter.project.LocalProjectInfo
 import com.intellij.ide.starter.runner.CurrentTestMethod
 import com.intellij.ide.starter.runner.Starter
 import org.junit.jupiter.api.AfterAll
+import org.junit.jupiter.api.Assertions.assertFalse
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
@@ -22,6 +23,9 @@ import org.kodein.di.DI
 import org.kodein.di.bindSingleton
 import software.aws.toolkits.jetbrains.uitests.TestCIServer
 import software.aws.toolkits.jetbrains.uitests.clearAwsXmlFile
+import software.aws.toolkits.jetbrains.uitests.docTests.prepTestData
+import software.aws.toolkits.jetbrains.uitests.docTests.scripts.updateReadmeLatestChangesConfirmOptionsTestScript
+import software.aws.toolkits.jetbrains.uitests.docTests.scripts.updateReadmeLatestChangesMakeChangesFlowTestScript
 import software.aws.toolkits.jetbrains.uitests.docTests.scripts.updateReadmeLatestChangesTestScript
 import software.aws.toolkits.jetbrains.uitests.executePuppeteerScript
 import software.aws.toolkits.jetbrains.uitests.setupTestEnvironment
@@ -49,6 +53,84 @@ class UpdateReadmeLatestChangesTest {
     fun setUp() {
         // Setup test environment
         setupTestEnvironment()
+        // prep test data - restore readme contents
+        prepTestData(false)
+    }
+
+    @Test
+    fun `Make Changes button leads to UPDATE with specific changes flow`() {
+        val testCase = TestCase(
+            IdeProductProvider.IC,
+            LocalProjectInfo(
+                Paths.get("tstData", "qdoc", "updateFlow")
+            )
+        ).useRelease(System.getProperty("org.gradle.project.ideProfileName"))
+
+        // inject connection
+        useExistingConnectionForTest()
+
+        Starter.newContext(CurrentTestMethod.hyphenateWithClass(), testCase).apply {
+            System.getProperty("ui.test.plugins").split(File.pathSeparator).forEach { path ->
+                pluginConfigurator.installPluginFromPath(
+                    Path.of(path)
+                )
+            }
+
+            copyExistingConfig(Paths.get("tstData", "configAmazonQTests"))
+            updateGeneralSettings()
+        }.runIdeWithDriver()
+            .useDriverAndCloseIde {
+                waitForProjectOpen()
+                // required wait time for the system to be fully ready
+                Thread.sleep(30000)
+
+                val result = executePuppeteerScript(updateReadmeLatestChangesMakeChangesFlowTestScript)
+
+                if (result.contains("Error: Test Failed")) {
+                    println("result: $result")
+                }
+
+                assertTrue(result.contains("Test Successful"))
+                assertFalse(result.contains("Error: Test Failed"))
+            }
+    }
+
+    @Test
+    fun `Prompted to confirm selected folder`() {
+        val testCase = TestCase(
+            IdeProductProvider.IC,
+            LocalProjectInfo(
+                Paths.get("tstData", "qdoc", "updateFlow")
+            )
+        ).useRelease(System.getProperty("org.gradle.project.ideProfileName"))
+
+        // inject connection
+        useExistingConnectionForTest()
+
+        Starter.newContext(CurrentTestMethod.hyphenateWithClass(), testCase).apply {
+            System.getProperty("ui.test.plugins").split(File.pathSeparator).forEach { path ->
+                pluginConfigurator.installPluginFromPath(
+                    Path.of(path)
+                )
+            }
+
+            copyExistingConfig(Paths.get("tstData", "configAmazonQTests"))
+            updateGeneralSettings()
+        }.runIdeWithDriver()
+            .useDriverAndCloseIde {
+                waitForProjectOpen()
+                // required wait time for the system to be fully ready
+                Thread.sleep(30000)
+
+                val result = executePuppeteerScript(updateReadmeLatestChangesConfirmOptionsTestScript)
+
+                if (result.contains("Error: Test Failed")) {
+                    println("result: $result")
+                }
+
+                assertTrue(result.contains("Test Successful"))
+                assertFalse(result.contains("Error: Test Failed"))
+            }
     }
 
     @Test
@@ -97,21 +179,6 @@ class UpdateReadmeLatestChangesTest {
         @AfterAll
         fun clearAwsXml() {
             clearAwsXmlFile()
-        }
-
-        @JvmStatic
-        @AfterAll
-        fun tearDown() {
-            val path = Paths.get("tstData", "qdoc", "updateFlow", "README.md").toUri()
-
-            val process = ProcessBuilder("git", "restore", path.path).start()
-            val exitCode = process.waitFor()
-            if (exitCode != 0) {
-                println("Warning: git stash command failed with exit code $exitCode")
-                process.errorStream.bufferedReader().use { reader ->
-                    println("Error: ${reader.readText()}")
-                }
-            }
         }
     }
 }
